+++ src/main/java/chess/domain/movement/Mobility.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.movement;
 
 public enum Mobility {
     ONCE,
