+++ src/main/java/chess/domain/movement/Movement.java
@@ -1,4 +1,6 @@
-package chess.domain;
+package chess.domain.movement;
+
+import chess.domain.position.RelativePosition;
 
 import java.util.List;
