+++ src/main/java/chess/view/CommandDto.java
@@ -1,6 +1,6 @@
 package chess.view;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 
 public class CommandDto {
